post and delete blog working

+++ blog/views.py
@@ -1,7 +1,7 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
 from rest_framework.permissions import IsAuthenticatedOrReadOnly
 
 from .models import Blog
@@ -18,6 +18,16 @@ class BlogListView(APIView):
         blogs = Blog.objects.all()
         serialized_blogs = BlogSerializer(blogs, many=True)
         return Response(serialized_blogs.data, status=status.HTTP_200_OK)
+
+    # Post Blog
+    def post(self, request):
+        request.data['owner'] = request.user.id
+        blog_to_add = BlogSerializer(data=request.data)
+        if blog_to_add.is_valid():
+            blog_to_add.save()
+            return Response(blog_to_add.data, status=status.HTTP_201_CREATED)
+        return Response(blog_to_add.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+
 
 
 class BlogDetailView(APIView):
@@ -37,8 +47,10 @@ class BlogDetailView(APIView):
         return Response(serialized_blog.data, status=status.HTTP_200_OK)
 
     # Delete Blog
-    def delete(self, _request, pk):
+    def delete(self, request, pk):
         blog_to_delete = Blog.objects.get(pk=pk)
+        if blog_to_delete.owner != request.user:
+            raise PermissionDenied(detail="Unauthorised")
         blog_to_delete.delete()
         raise Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -52,10 +64,4 @@ class BlogDetailView(APIView):
             return Response(updated_blog.data, status=status.HTTP_202_ACCEPTED)
         return Response(updated_blog.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
 
-    # Post Blog
-    def post(self, request):
-        blog_to_add = BlogSerializer(data=request.data)
-        if blog_to_add.is_valid():
-            blog_to_add.save()
-            return Response(blog_to_add.data, status=status.HTTP_201_CREATED)
-        return Response(blog_to_add.errors, status=status.HTTP_422_UNPROCESSABLE_ENTITY)
+
